requirements.txt
atomicwrites==1.4.0
attrs==21.2.0
backports.csv==1.0.7
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.7
cheroot==8.5.2
CherryPy==18.6.1
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
decorator==4.4.2
feedparser==6.0.8
future==0.18.2
huepy==1.2.1
idna==3.3
imageio==2.9.0
imageio-ffmpeg==0.4.5
iniconfig==1.1.1
instabot==0.117.0
jaraco.classes==3.2.1
jaraco.collections==3.4.0
jaraco.functools==3.3.0
jaraco.text==3.5.1
joblib==1.1.0
lxml==4.6.3
mock==4.0.3
more-itertools==8.10.0
moviepy==1.0.3
mysqlclient==2.0.3
nltk==3.6.5
numpy==1.21.2
packaging==21.0
Pattern==3.6
pdfminer.six==20211012
Pillow==8.3.2
pluggy==1.0.0
portend==3.0.0
proglog==0.1.9
py==1.10.0
pycparser==2.20
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.5
python-docx==0.8.11
pytz==2021.3
pywin32==302
regex==2021.10.8
requests==2.26.0
requests-toolbelt==0.9.1
responses==0.14.0
schedule==1.1.0
scipy==1.7.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.2.1
tempora==4.1.2
toml==0.10.2
tqdm==4.62.3
urllib3==1.26.7
zc.lockfile==2.0
